refactor reporting job routes
