+++ src/app/routers/reporting_jobs.py
@@ -1,0 +1,324 @@
+from typing import Annotated
+
+from fastapi import APIRouter, Header, Path, Query
+
+from app.contracts.reporting import (
+    REPORT_JOB_LIST_RESPONSE_EXAMPLE,
+    ReportJobListResponse,
+    ReportJobStatusEventsResponse,
+    ReportJobStatusResponse,
+    ReportSnapshotLineageResponse,
+)
+from app.middleware.correlation import correlation_id_var
+from app.routers.reporting_context import reporting_context_headers
+from app.routers.reporting_errors import report_job_error_response
+from app.services.reporting_service_provider import reporting_job_query_service
+
+jobs_router = APIRouter(prefix="/api/v1/report-jobs", tags=["Report Jobs"])
+
+
+@jobs_router.get(
+    "",
+    response_model=ReportJobListResponse,
+    summary="Search report jobs for operations and support",
+    description=(
+        "Return a bounded list of report jobs through the governed gateway boundary. Use this "
+        "endpoint when operators or support tooling need to find jobs by tenant, region, status, "
+        "portfolio, as-of date, idempotency key, or correlation identifier before drilling into "
+        "status or event history."
+    ),
+    openapi_extra={
+        "responses": {
+            "200": {
+                "content": {
+                    "application/json": {
+                        "example": REPORT_JOB_LIST_RESPONSE_EXAMPLE,
+                    }
+                }
+            }
+        }
+    },
+    responses={
+        **report_job_error_response(
+            400,
+            example_key="invalid_report_job_filters",
+            description="Returned when no supported job-search filter is supplied.",
+        ),
+        **report_job_error_response(
+            502,
+            example_key="report_job_upstream_unavailable",
+            description="Returned when lotus-report is unavailable or returns an unsafe failure.",
+        ),
+    },
+)
+async def list_report_jobs(
+    tenant_id_filter: Annotated[
+        str | None,
+        Query(alias="tenantId", description="Return only jobs for this tenant identifier."),
+    ] = None,
+    region_filter: Annotated[
+        str | None,
+        Query(alias="region", description="Return only jobs for this operating region."),
+    ] = None,
+    status_filter: Annotated[
+        str | None,
+        Query(alias="status", description="Return only jobs in this current lifecycle status."),
+    ] = None,
+    report_type_filter: Annotated[
+        str | None,
+        Query(alias="reportType", description="Return only jobs for this report type."),
+    ] = None,
+    portfolio_id_filter: Annotated[
+        str | None,
+        Query(
+            alias="portfolioId",
+            description="Return only jobs whose scope includes this portfolio.",
+        ),
+    ] = None,
+    as_of_date_filter: Annotated[
+        str | None,
+        Query(alias="asOfDate", description="Return only jobs for this business as-of date."),
+    ] = None,
+    idempotency_key_filter: Annotated[
+        str | None,
+        Query(alias="idempotencyKey", description="Return only jobs for this idempotency key."),
+    ] = None,
+    correlation_id_filter: Annotated[
+        str | None,
+        Query(
+            alias="correlationId",
+            description="Return only jobs for this correlation identifier.",
+        ),
+    ] = None,
+    created_from: Annotated[
+        str | None,
+        Query(alias="createdFrom", description="Inclusive UTC lower bound for job creation time."),
+    ] = None,
+    created_to: Annotated[
+        str | None,
+        Query(alias="createdTo", description="Inclusive UTC upper bound for job creation time."),
+    ] = None,
+    limit: Annotated[
+        int,
+        Query(
+            alias="limit",
+            ge=1,
+            le=100,
+            description="Maximum number of report jobs returned by this bounded search.",
+        ),
+    ] = 25,
+    actor_id: Annotated[str | None, Header(alias="X-Actor-Id")] = None,
+    caller_application: Annotated[str | None, Header(alias="X-Caller-Application")] = None,
+    tenant_id: Annotated[str | None, Header(alias="X-Tenant-Id")] = None,
+    region: Annotated[str | None, Header(alias="X-Region")] = None,
+    booking_center_code: Annotated[str | None, Header(alias="X-Booking-Center-Code")] = None,
+    role: Annotated[str | None, Header(alias="X-Role")] = None,
+) -> ReportJobListResponse:
+    filters = {
+        "tenantId": tenant_id_filter,
+        "region": region_filter,
+        "status": status_filter,
+        "reportType": report_type_filter,
+        "portfolioId": portfolio_id_filter,
+        "asOfDate": as_of_date_filter,
+        "idempotencyKey": idempotency_key_filter,
+        "correlationId": correlation_id_filter,
+        "createdFrom": created_from,
+        "createdTo": created_to,
+        "limit": limit,
+    }
+    filters = {key: value for key, value in filters.items() if value is not None}
+    return await reporting_job_query_service().list_report_jobs(
+        filters=filters,
+        caller_headers=reporting_context_headers(
+            actor_id=actor_id,
+            caller_application=caller_application,
+            tenant_id=tenant_id,
+            region=region,
+            booking_center_code=booking_center_code,
+            role=role,
+        ),
+        correlation_id=correlation_id_var.get(),
+    )
+
+
+@jobs_router.get(
+    "/{job_id}",
+    response_model=ReportJobStatusResponse,
+    summary="Get report job status",
+    description=(
+        "Return product-safe report job status and diagnostics from lotus-report. Use this "
+        "endpoint after submit or search when a caller needs current lifecycle state for one job."
+    ),
+    responses={
+        **report_job_error_response(
+            404,
+            example_key="report_job_not_found",
+            description="Returned when the requested report job does not exist.",
+        ),
+        **report_job_error_response(
+            502,
+            example_key="report_job_upstream_unavailable",
+            description="Returned when lotus-report is unavailable or returns an unsafe failure.",
+        ),
+    },
+)
+async def get_report_job_status(
+    job_id: Annotated[str, Path(description="Opaque report job identifier.")],
+    actor_id: Annotated[str | None, Header(alias="X-Actor-Id")] = None,
+    caller_application: Annotated[str | None, Header(alias="X-Caller-Application")] = None,
+    tenant_id: Annotated[str | None, Header(alias="X-Tenant-Id")] = None,
+    region: Annotated[str | None, Header(alias="X-Region")] = None,
+    booking_center_code: Annotated[str | None, Header(alias="X-Booking-Center-Code")] = None,
+    role: Annotated[str | None, Header(alias="X-Role")] = None,
+) -> ReportJobStatusResponse:
+    return await reporting_job_query_service().get_report_job_status(
+        job_id=job_id,
+        caller_headers=reporting_context_headers(
+            actor_id=actor_id,
+            caller_application=caller_application,
+            tenant_id=tenant_id,
+            region=region,
+            booking_center_code=booking_center_code,
+            role=role,
+        ),
+        correlation_id=correlation_id_var.get(),
+    )
+
+
+@jobs_router.get(
+    "/{job_id}/events",
+    response_model=ReportJobStatusEventsResponse,
+    summary="Get report job event history",
+    description=(
+        "Return append-only report job lifecycle events through the governed gateway boundary. "
+        "Use this endpoint for operational support when current status alone is insufficient."
+    ),
+    responses={
+        **report_job_error_response(
+            404,
+            example_key="report_job_not_found",
+            description="Returned when the requested report job does not exist.",
+        ),
+        **report_job_error_response(
+            502,
+            example_key="report_job_upstream_unavailable",
+            description="Returned when lotus-report is unavailable or returns an unsafe failure.",
+        ),
+    },
+)
+async def get_report_job_events(
+    job_id: Annotated[str, Path(description="Opaque report job identifier.")],
+    actor_id: Annotated[str | None, Header(alias="X-Actor-Id")] = None,
+    caller_application: Annotated[str | None, Header(alias="X-Caller-Application")] = None,
+    tenant_id: Annotated[str | None, Header(alias="X-Tenant-Id")] = None,
+    region: Annotated[str | None, Header(alias="X-Region")] = None,
+    booking_center_code: Annotated[str | None, Header(alias="X-Booking-Center-Code")] = None,
+    role: Annotated[str | None, Header(alias="X-Role")] = None,
+) -> ReportJobStatusEventsResponse:
+    return await reporting_job_query_service().get_report_job_events(
+        job_id=job_id,
+        caller_headers=reporting_context_headers(
+            actor_id=actor_id,
+            caller_application=caller_application,
+            tenant_id=tenant_id,
+            region=region,
+            booking_center_code=booking_center_code,
+            role=role,
+        ),
+        correlation_id=correlation_id_var.get(),
+    )
+
+
+@jobs_router.get(
+    "/{job_id}/lineage",
+    response_model=ReportSnapshotLineageResponse,
+    summary="Get report snapshot lineage",
+    description=(
+        "Return lineage evidence for a report job's captured input snapshot and upstream "
+        "dependency calls through the governed gateway boundary."
+    ),
+    responses={
+        **report_job_error_response(
+            404,
+            example_key="report_snapshot_not_found",
+            description="Returned when snapshot lineage is unavailable for this report job.",
+        ),
+        **report_job_error_response(
+            502,
+            example_key="report_job_upstream_unavailable",
+            description="Returned when lotus-report is unavailable or returns an unsafe failure.",
+        ),
+    },
+)
+async def get_report_job_lineage(
+    job_id: Annotated[str, Path(description="Opaque report job identifier.")],
+    actor_id: Annotated[str | None, Header(alias="X-Actor-Id")] = None,
+    caller_application: Annotated[str | None, Header(alias="X-Caller-Application")] = None,
+    tenant_id: Annotated[str | None, Header(alias="X-Tenant-Id")] = None,
+    region: Annotated[str | None, Header(alias="X-Region")] = None,
+    booking_center_code: Annotated[str | None, Header(alias="X-Booking-Center-Code")] = None,
+    role: Annotated[str | None, Header(alias="X-Role")] = None,
+) -> ReportSnapshotLineageResponse:
+    return await reporting_job_query_service().get_report_job_lineage(
+        job_id=job_id,
+        caller_headers=reporting_context_headers(
+            actor_id=actor_id,
+            caller_application=caller_application,
+            tenant_id=tenant_id,
+            region=region,
+            booking_center_code=booking_center_code,
+            role=role,
+        ),
+        correlation_id=correlation_id_var.get(),
+    )
+
+
+@jobs_router.post(
+    "/{job_id}/cancel",
+    response_model=ReportJobStatusResponse,
+    summary="Cancel report job before render or archive",
+    description=(
+        "Cancel a report job while it is still before render, archive, or completion. Use this "
+        "endpoint only for bounded pre-render cancellation; rerender, reissue, archive, and legal "
+        "hold semantics are owned by later reporting RFCs."
+    ),
+    responses={
+        **report_job_error_response(
+            404,
+            example_key="report_job_not_found",
+            description="Returned when the requested report job does not exist.",
+        ),
+        **report_job_error_response(
+            409,
+            example_key="report_job_cannot_be_cancelled",
+            description="Returned when the job has completed or was already cancelled.",
+        ),
+        **report_job_error_response(
+            502,
+            example_key="report_job_upstream_unavailable",
+            description="Returned when lotus-report is unavailable or returns an unsafe failure.",
+        ),
+    },
+)
+async def cancel_report_job(
+    job_id: Annotated[str, Path(description="Opaque report job identifier.")],
+    actor_id: Annotated[str | None, Header(alias="X-Actor-Id")] = None,
+    caller_application: Annotated[str | None, Header(alias="X-Caller-Application")] = None,
+    tenant_id: Annotated[str | None, Header(alias="X-Tenant-Id")] = None,
+    region: Annotated[str | None, Header(alias="X-Region")] = None,
+    booking_center_code: Annotated[str | None, Header(alias="X-Booking-Center-Code")] = None,
+    role: Annotated[str | None, Header(alias="X-Role")] = None,
+) -> ReportJobStatusResponse:
+    return await reporting_job_query_service().cancel_report_job(
+        job_id=job_id,
+        caller_headers=reporting_context_headers(
+            actor_id=actor_id,
+            caller_application=caller_application,
+            tenant_id=tenant_id,
+            region=region,
+            booking_center_code=booking_center_code,
+            role=role,
+        ),
+        correlation_id=correlation_id_var.get(),
+    )
